### 2024-09-04 #7180 @editorialbot
[PRE REVIEW]: shipgrav: A Python package for marine gravimetry
**Submitting author:** @hfmark (<a href="http://orcid.org/0000-0002-1722-3759">Hannah Mark</a>)
**Repository:** https://github.com/PFPE/shipgrav
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
