### 2024-09-04 d71aeab
url to doi for Eotvos
