+++ joss/paper.bib
@@ -145,7 +145,7 @@
 	title = {Experimenteller {Nachweis} der {Schwereänderung}, die ein auf normal geformter {Erdoberfläche} in östlicher oder westlicher {Richtung} bewegter {Körper} durch diese {Bewegung} erleidet},
 	volume = {364},
 	issn = {0003-3804},
-	url = {https://doi.org/10.1002/andp.19193641604},
+	doi = {10.1002/andp.19193641604},
 	language = {en},
 	number = {16},
 	urldate = {2024-07-17},
