### 2024-09-11 #7180 @kthyng
@hfmark some questions/comments:
* Could this be released onto PyPI or conda-forge so users could install that way? This wouldn't hold up review but I would like to see this happen as part of the review process, if possible.
* Say hi to Masako for me!!!
