+++ .github/workflows/release-publish.yml
@@ -27,27 +27,27 @@ jobs:
         name: python-package-distributions
         path: dist/
 
-#  publish-to-pypi:
-#    name: >-
-#      Publish Python 🐍 distribution 📦 to TestPyPI
-#    if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes
-#    needs:
-#    - build
-#    runs-on: ubuntu-latest
-#    environment:
-#      name: pypi
-#      url: https://pypi.org/p/<package-name>  # Replace <package-name> with your PyPI project name
-#    permissions:
-#      id-token: write  # IMPORTANT: mandatory for trusted publishing
-#
-#    steps:
-#    - name: Download all the dists
-#      uses: actions/download-artifact@v4
-#      with:
-#        name: python-package-distributions
-#        path: dist/
-#    - name: Publish distribution 📦 to PyPI
-#      uses: pypa/gh-action-pypi-publish@release/v1
+  publish-to-pypi:
+    name: >-
+      Publish Python 🐍 distribution 📦 to TestPyPI
+    if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes
+    needs:
+    - build
+    runs-on: ubuntu-latest
+    environment:
+      name: pypi
+      url: https://pypi.org/p/shipgrav
+    permissions:
+      id-token: write  # IMPORTANT: mandatory for trusted publishing
+
+    steps:
+    - name: Download all the dists
+      uses: actions/download-artifact@v4
+      with:
+        name: python-package-distributions
+        path: dist/
+    - name: Publish distribution 📦 to PyPI
+      uses: pypa/gh-action-pypi-publish@release/v1
 
   github-release:
     name: >-
