### 2024-09-17 c7e02a5
update docs for PyPI, matplotlib optional dependency, increment patch version
