### 2024-09-17 #7180 @hfmark
Hi @kthyng - I think I've sorted out pypi publishing, `pip install shipgrav` should work now! And the docs have been updated accordingly.

Also, a few more possibilities for reviewers that I don't think are in your database: GTAIto and dsandwell may or may not be big github users (and I don't know if they've ever done code review in this style) but both know a ton about gravity processing. jrleeman might also be a good option as a geophysics generalist (and looking at his profile I see he's written some gravity-related stuff)
