### 2024-10-04 #7180 @rwegener2
👋🏻 @akinremisa and @andreww, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
