### 2024-10-04 #7180 @andreww
@rwegener2 - sure, I'm happy to review this. I'm not a gravity person but I really should know enough to be useful (and it's mostly about the software). It won't be until late October though (start of term chaos here for a couple of weeks).
