### 2024-10-09 #7180 @rwegener2
👋🏻 @malmans2, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
