### 2024-10-25 PFPE/shipgrav#1 @andreww
Deprication warnings from the test cases
As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7358) I ran the test cases (with a conda-forge Python 3.9.20 and pandas version 2.2.3). The tests all pass, but I see the following warnings:

```
/Users/andreww/miniforge3/envs/shipgrav2/lib/python3.9/site-packages/shipgrav/io.py:470: FutureWarning: Support for nested sequences for 'parse_dates' in pd.read_csv is deprecated. Combine the desired columns with pd.to_datetime after parsing instead.
  dat = pd.read_csv(path, delimiter=',', names=['date', 'time', 'counts'],
./Users/andreww/miniforge3/envs/shipgrav2/lib/python3.9/site-packages/shipgrav/io.py:385: FutureWarning: Support for nested sequences for 'parse_dates' in pd.read_csv is deprecated. Combine the desired columns with pd.to_datetime after parsing instead.
  dat = pd.read_csv(path, delimiter=' ', names=['date', 'time', 'grav', 'lat', 'lon'],
.../Users/andreww/miniforge3/envs/shipgrav2/lib/python3.9/site-packages/shipgrav/io.py:749: FutureWarning: errors='ignore' is deprecated and will raise in a future version. Use to_numeric without passing `errors` and catch exceptions explicitly instead
  return df.apply(pd.to_numeric, errors='ignore'), col_info
```

At some point it's probably worth updating those three bits of code to avoid things falling over in the future.
