### 2024-10-25 PFPE/shipgrav#2 @andreww
Tests fail under python 3.13
I tried running the tests as part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7358. I see errors like this when trying to run the test cases under python 3.13 (i.e. the version currently installed by default by conda).

```
________________ ERROR collecting shipgrav/tests/test_grav_data.py _________________
ImportError while importing test module '/Users/andreww/Code/shipgrav/shipgrav/shipgrav/tests/test_grav_data.py'.
Traceback:
../../../../miniforge3/envs/shipgrav/lib/python3.13/importlib/__init__.py:88: in import_module
tests/__init__.py:5: in <module>
    from pkg_resources import resource_filename
E   ModuleNotFoundError: No module named 'pkg_resources'
```

I think this is basically the same as the error described at https://github.com/mu-editor/mu/issues/2485 - I think the links in that issue describe the fixes.
