### 2024-10-25 PFPE/shipgrav#3 @andreww
Test automation
This is also part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7358.

It would be a really good idea if the tests were set up to run in an automated manner (on multiple python versions) as and when the code is changed. One way to do this would be to plug the tests into github actions with the appropriate triggers. Some documentation can be found at: https://docs.github.com/en/actions/use-cases-and-examples/building-and-testing/building-and-testing-python - and if it's useful I can probably find a relevant example case without too much effort.
