### 2024-10-25 PFPE/shipgrav#4 @andreww
Contributors
This part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7358 where I'm asked "Has the submitting author (@hfmark) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?"

I'm happy to say yes to the first question, but it would be useful to know what the contributions of the other authors is before confirming the second. Currently all commits are made by @hfmark and only one name appears in https://github.com/PFPE/shipgrav/blob/main/CONTRIBUTORS. It looks like there is much history of this package prior to the first github upload in January this year. Maybe the best approach is to briefly document the earlier history and contributions to the development of the package (which could be supervision/leadership/data/testing/expertise rather than just lines of code) in the documentation and update the CONTRIBUTORS file.
