### 2024-10-25 PFPE/shipgrav#5 @andreww
Prominent link to documentation
This is a super minor thing that I noted doing the JOSS review: openjournals/joss-reviews#7358

It's probably worth adding a prominent link to the documentation at the top of the github repo as well as in the readme file (using the little cog next to "about").
