### 2024-10-25 PFPE/shipgrav#6 @andreww
Some example scripts don't seem to run
This is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7358

After running the download_data.sh script (and seeing data being downloaded and being unpacked), I tried running `python dgs_bgm_comp.py` to verify the examples run. The script fails with:

```
Traceback (most recent call last):
  File "/Users/andreww/Code/shipgrav_2/shipgrav/example-scripts/dgs_bgm_comp.py", line 74, in <module>
    bgm_data = sgi.read_bgm_raw(bgm_files, ship)
  File "/Users/andreww/miniforge3/envs/shipgrav2/lib/python3.9/site-packages/shipgrav/io.py", line 450, in read_bgm_raw
    return pd.concat(dats, ignore_index=True)
  File "/Users/andreww/miniforge3/envs/shipgrav2/lib/python3.9/site-packages/pandas/core/reshape/concat.py", line 382, in concat
    op = _Concatenator(
  File "/Users/andreww/miniforge3/envs/shipgrav2/lib/python3.9/site-packages/pandas/core/reshape/concat.py", line 445, in __init__
    objs, keys = self._clean_keys_and_objs(objs, keys)
  File "/Users/andreww/miniforge3/envs/shipgrav2/lib/python3.9/site-packages/pandas/core/reshape/concat.py", line 507, in _clean_keys_and_objs
    raise ValueError("No objects to concatenate")
ValueError: No objects to concatenate
```
because `ship` is an empty list. Probably a issue with the downloaded directory structure and the globbing I expect. It may be more reliable to use something like the pooch package to manage the data download for the examples.
