### 2024-10-25 #7358 @andreww
I've now had a chance to look over and try out the code, and read the manuscript. The manuscript itself looks good and is easy to read. One minor thing is that the role of the contributors could be clarified (see https://github.com/PFPE/shipgrav/issues/4).

I ran into problems running the tests (see https://github.com/PFPE/shipgrav/issues/1, https://github.com/PFPE/shipgrav/issues/2 and https://github.com/PFPE/shipgrav/issues/3) and the examples (https://github.com/PFPE/shipgrav/issues/6). I expect this could all be fixed by automating the tests (and covering different python versions) and making the documentation around setting up and running the examples a bit clearer / more reproducible.

Not in an issue, but it would be useful to add some guidelines for (new) contributors, and maybe look at the paper and think about giving more explicit information on the "state of the field" (there is some info in there around line 31 and 32, but are there any comparable packages for parts of the pipe line?)
