### 2024-10-25 #7358 @hfmark
Thanks @andreww ! Most of the issues you've raised I know I can fix easily; the thing with the example scripts running looks like it's a problem with the r2r data download. `curl` used to work! But trying it again on my own computer, it's failing to download most of the files, hence the empty list of BGM filenames :frowning_face:. I'll work on all of the above and update on the issues as I go.

As for comparable packages, I can think of two that are public that cover some bits of this workflow, though not the whole thing. Not that marine gravity specialists are especially secretive, but I do think we've been a bit behind the curve in terms of open source. Fair point about expanding on that in the paper, I'll work on that as well.
