### 2024-10-25 ddd3770
update from pkg_resources to importlib.resources, see issue #2
