+++ shipgrav/tests/__init__.py
@@ -2,15 +2,16 @@
 Tests for shipgrav
 """
 
-from pkg_resources import resource_filename
 import sys
 import unittest
+import importlib.resources as importlib_resources
 
 
 def run():
     loader = unittest.TestLoader()
-    test_dir = resource_filename('shipgrav', 'tests')
-    suite = loader.discover(test_dir)
+    ref = importlib_resources.files('shipgrav') / 'tests'
+    with importlib_resources.as_file(ref) as path:
+        suite = loader.discover(ref)
     runner = unittest.runner.TextTestRunner()  # verbosity=2)
     ret = not runner.run(suite).wasSuccessful()
     sys.exit(ret)
