### 2024-10-25 eb9d417
fix DeprecationWarnings per issue #1, and also SyntaxWarnings for regex
