+++ shipgrav/io.py
@@ -169,7 +169,7 @@ def _navdate_Atlantis(allnav, talker):
 
     for i in range(N):
         pre = subnav[i].split(talker)[0]
-        date = re.findall('NAV (\d{4})/(\d{2})/(\d{2})', pre)[0]
+        date = re.findall(r'NAV (\d{4})/(\d{2})/(\d{2})', pre)[0]
         year = int(date[0])  # year
         mon = int(date[1])  # month
         day = int(date[2])  # day
@@ -191,7 +191,7 @@ def _navdate_NBP(allnav, talker):
 
     for i in range(N):
         pre = subnav[i].split(talker)[0]
-        date = re.findall('(\d{2})\+(\d{2,3}):.*', pre)[0]
+        date = re.findall(r'(\d{2})\+(\d{2,3}):.*', pre)[0]
         # year (NBP didn't exist before 2000 so this is ok)
         year = '20' + date[0]
         doy = date[1]  # doy
@@ -216,7 +216,7 @@ def _navdate_Thompson(allnav, talker):
 
     for i in range(N):
         pre = subnav[i].split(talker)[0]
-        date = re.findall('(\d{2})/(\d{2})/(\d{4}),*', pre)[0]
+        date = re.findall(r'(\d{2})/(\d{2})/(\d{4}),*', pre)[0]
         year = int(date[2])
         mon = int(date[0])
         day = int(date[1])
@@ -247,8 +247,8 @@ def _navdate_Revelle(allnav, talker):
         for k in range(inds[i]-1, j, -1):  # step backwards toward the last talker line
             before = allnav[k]
             # date is at the start of this line
-            if re.match('(\d{4})-(\d{2})-(\d{2})T*', before):
-                date = re.findall('(\d{4})-(\d{2})-(\d{2})T*', before)[0]
+            if re.match(r'(\d{4})-(\d{2})-(\d{2})T*', before):
+                date = re.findall(r'(\d{4})-(\d{2})-(\d{2})T*', before)[0]
                 year = int(date[0])
                 mon = int(date[1])
                 day = int(date[2])
@@ -269,12 +269,12 @@ def _navdate_Ride(allnav, talker):
 
     for i in range(N):
         if talker == 'INGGA':  # on Ride, uses posix timestamps
-            date = re.findall('(\d+(\.\d*)?) \$%s' % talker, subnav[i])[0]
+            date = re.findall(r'(\d+(\.\d*)?) \$%s' % talker, subnav[i])[0]
             timest[i] = datetime.fromtimestamp(
                 float(date[0]), tzinfo=timezone.utc)
         elif talker == 'GPGGA':  # includes time only with date, unlike other GPGGAs
             date = re.findall(
-                '(\d{4})\-(\d{2})\-(\d{2})T(\d{2}):(\d{2}):(\d{2})\.(\d.*?)Z', subnav[i])[0]
+                r'(\d{4})\-(\d{2})\-(\d{2})T(\d{2}):(\d{2}):(\d{2})\.(\d.*?)Z', subnav[i])[0]
             year = int(date[0])
             mon = int(date[1])
             day = int(date[2])
@@ -383,9 +383,8 @@ def read_bgm_rgs(fp, ship):
     dats = []
     for path in fp:
         dat = pd.read_csv(path, delimiter=' ', names=['date', 'time', 'grav', 'lat', 'lon'],
-                          usecols=(1, 2, 3, 11, 12), parse_dates=[[0, 1]])
-        ndt = [e.tz_localize(timezone.utc) for e in dat['date_time']]
-        dat['date_time'] = ndt
+                          usecols=(1, 2, 3, 11, 12))
+        dat['date_time'] = pd.to_datetime(dat.pop('date')+' '+dat.pop('time'),utc=True)
         dats.append(dat)
 
[…]
