### 2024-10-25 PFPE/shipgrav#6 @hfmark
The R2R links were failing with `curl` because something changed on the server side, I think. Adding -L fixes the download process, which should solve this issue.
