### 2024-10-25 PFPE/shipgrav#1 @hfmark
Date parsing is updated for all the cases that were using nested sequences, and the to_numeric() process now goes column by column so that datetime strings are preserved without errors. I also fixed some syntax warnings that I saw when I tried 3.13, for escaping in regex, and tests seem to be warning-free.
