### 2024-10-25 7b768af
update test suite per issue #3
