+++ .github/workflows/test.yml
@@ -1,7 +1,7 @@
 name: tests
 on:
-  #push:
-  workflow_dispatch:
+  push:
+  #workflow_dispatch:
   pull_request:
     branches:
       # only branches from forks which have the form 'user:branch-name'
@@ -13,29 +13,10 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest]
-        python: [3.9, "3.10"]
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python: [3.9, "3.10","3.13"]
         add_package: [""]
         continue-on-error: [false]
-        include:
-          - os: macos-latest
-            python: 3.9
-            continue-on-error: false
-          - os: windows-latest
-            python: 3.9
-            continue-on-error: false
-          - os: ubuntu-latest
-            python: 3.9
-            add_package: features
-            continue-on-error: true
-          - os: macos-latest
-            python: 3.9
-            add_package: features
-            continue-on-error: true
-          - os: windows-latest
-            python: 3.9
-            add_package: features
-            continue-on-error: true
     runs-on: ${{ matrix.os }}
     continue-on-error: ${{ matrix.continue-on-error }}
     defaults:
@@ -63,6 +44,6 @@ jobs:
           coverage run --include='*/shipgrav/*' -m shipgrav.tests
           coverage xml -o ../coverage.xml
       - name: upload coverage
-        uses: codecov/codecov-action@v3
+        uses: codecov/codecov-action@v4
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
