### 2024-10-25 PFPE/shipgrav#3 @hfmark
I had them set with a cron job; that's still there, but now they run on push as well. I also cleaned up the os matrix and added 3.13, which passes :slightly_smiling_face:
