### 2024-10-25 aed66bd
update contributors file with more detailed info, see issue #4
