### 2024-10-25 PFPE/shipgrav#4 @hfmark
The CRediT taxonomy maybe isn't designed for code, but it's a nice standard, so I used its definitions to update the CONTRIBUTORS file with more details on who did what.
