### 2024-10-25 0b1f18a
add a sentence about other software options
