### 2024-10-26 dd56cce
fix chained assignment for embedding track in satellite data
