### 2024-10-26 2a01bf7
update contributor guidelines
