### 2024-10-26 71429ef
bump version, and update optional dependency list
