### 2024-10-26 #7358 @hfmark
ok, I think I got all the issues raised so far (and a few extras that popped up along the way)! I updated the proof above for the one slight change to the paper.
