### 2024-10-28 PFPE/shipgrav#10 @malmans2
Improve test coverage
PR #9 allows running the test suite from the root directory, making it easier to execute tests and use different testing tools. I ran pytest and pytest-cov, and noticed that the io and grav module have lower coverage.

Is it possible to improve the coverage?

After cloning my branch from PR #9, you can reproduce what I did with the snippet below:
```
cd shipgrav
pip install pytest pytest-cov
pytest -vv --cov=. .
```
```
--------- coverage: platform darwin, python 3.11.10-final-0 ----------
Name                                 Stmts   Miss  Cover
--------------------------------------------------------
shipgrav/__init__.py                     0      0   100%
shipgrav/grav.py                       491    113    77%
shipgrav/io.py                         385    166    57%
shipgrav/nav.py                         30      0   100%
shipgrav/tests/__init__.py              11      7    36%
shipgrav/tests/__main__.py               2      2     0%
shipgrav/tests/test_grav_data.py        47      2    96%
shipgrav/tests/test_grav_nodata.py      30      2    93%
shipgrav/tests/test_io.py               38      2    95%
shipgrav/tests/test_nav.py              28      2    93%
shipgrav/tests/test_utils.py            21      2    90%
shipgrav/utils.py                      137     93    32%
--------------------------------------------------------
TOTAL                                 1220    391    68%
```

openjournals/joss-reviews#7358
