### 2024-10-28 PFPE/shipgrav#11 @malmans2
New PRs do not trigger CI
When I opened PR #9, I noticed the CI wasn't triggered. This is controlled by the YAML file here: https://github.com/PFPE/shipgrav/blob/main/.github/workflows/test.yml

Is this intentional? Wouldn't it be better to configure something like this?
```yaml
on:
  push:
    branches:
    - main
  pull_request:
    branches:
    - main
```

openjournals/joss-reviews#7358
