### 2024-10-28 PFPE/shipgrav#12 @malmans2
Follow PEP8 standards
Especially if you encourage external contributions, I suggest following [PEP8](https://peps.python.org/pep-0008/) standards as closely as possible. There are many linters and formatters that can help easily check and enforce these standards. For example, I recommend running:

- [isort](https://pycqa.github.io/isort/): sort imports
- [black](https://github.com/psf/black): code formatter
- [flake8](https://flake8.pycqa.org/en/latest/): enforce the PEP8 style guide

openjournals/joss-reviews#7358
