### 2024-10-28 #7358 @malmans2
Hi all,

I've started the review and opened a few issues/PRs. I'm not finished yet, but I should be able to complete it this week. Once I'm done, I'll post a recap here with any outstanding issues (if there are any).

In the meantime, @hfmark, could you please clarify the contributions of your co-authors? We need some clarification because you're the only contributor to the code on GitHub.

You can find JOSS authorship policy here: https://joss.readthedocs.io/en/latest/submitting.html#authorship

Thanks!
