### 2024-10-28 #7358 @malmans2
> so if you pip update you'll get at least one of your issues addressed (CI, tests run on push).

I started the review using the latest version, but the CI on GitHub isn’t affected by the PyPI version. You may want to adjust the CI workflow configuration. I’ve left more details in the issue: https://github.com/PFPE/shipgrav/issues/11
