### 2024-10-28 #7358 @hfmark
> the CI on GitHub isn’t affected by the PyPI version

yep, sorry, I shouldn't have written that reply before coffee this morning! I did update the tests workflow, but I'll address it in the issue since the updated version is still set up slightly differently from your suggestion.
