+++ example-scripts/dgs_bgm_comp.py
@@ -1,12 +1,10 @@
 import os
-import sys
 from glob import glob
 
 import matplotlib.pyplot as plt
 import numpy as np
 import shipgrav.grav as sgg
 import shipgrav.io as sgi
-import shipgrav.nav as sgn
 import tomli as tm
 from scipy.interpolate import interp1d
 from scipy.signal import filtfilt, firwin
