+++ example-scripts/dgs_ccp_calc.py
@@ -1,5 +1,4 @@
 import os
-import sys
 from glob import glob
 
 import matplotlib.pyplot as plt
