+++ example-scripts/RMBA_calc.py
@@ -1,6 +1,4 @@
 import os
-import sys
-from glob import glob
 
 import matplotlib.pyplot as plt
 import numpy as np
