+++ example-scripts/interactive_line_pick.py
@@ -1,5 +1,4 @@
 import os
-import sys
 from glob import glob
 from shutil import rmtree
 
@@ -7,10 +6,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 import shipgrav.grav as sgg
 import shipgrav.io as sgi
-import shipgrav.nav as sgn
 import shipgrav.utils as sgu
 import tomli as tm
-from matplotlib.backend_bases import MouseEvent
 from matplotlib.gridspec import GridSpec
 from scipy.interpolate import interp1d
 from scipy.signal import filtfilt, firwin
