### 2024-10-30 PFPE/shipgrav#12 @hfmark
I did run `autopep8` on everything, but I guess that only covers some aspects. a4c8a60f3e0dc8a26aa3da9abadba8c314dbec86 and 6eefc086f82d4f0c92168429e5df80efe1363b25 are running `isort` and some corrections from `flake8`. Personally I don't love the "blackened" style - I find it less readable - but if it's really important I can add it.
