### 2024-10-30 a0f7a68
use system path separator so windows tests can run, bug from PR #9
