### 2024-10-30 c16c16f
update test triggers per issue #11
