### 2024-10-30 PFPE/shipgrav#12 @malmans2
Nowadays, it's hard to find open-source software that doesn't use Black or Ruff. In my opinion, this improves readability, as it's the standard most developers are familiar with. However, code formatting is not a requirement for JOSS, so it's not mandatory.
