### 2024-10-30 PFPE/shipgrav#13 @malmans2
Example scripts feedback
I executed the example scripts, and they work great. I have a few suggestions that could enhance the user experience:

- It would be helpful to include these example scripts directly in the documentation. For instance, you could embed the code within a Jupyter notebook, allowing users to easily follow along with the code and view the figures simultaneously. Sphinx supports notebook rendering. See: https://docs.readthedocs.io/en/stable/guides/jupyter.html
- Some scripts take a bit of time to run (e.g., `mru_coherence.py`), and it was unclear whether they were stuck or simply processing. I suggest adding logs and/or progress bars, either within the main code or in the example scripts.
- The bash script for downloading data works well, but why not use Python for consistency with the rest of the codebase? Tools like Pooch could be used for this, allowing you to implement a download module in the codebase to be used by the example scripts. See: https://www.fatiando.org/pooch/latest/

openjournals/joss-reviews#7358
