### 2024-10-30 PFPE/shipgrav#14 @malmans2
Mention how to install optional dependencies
I suggest explicitly mentioning in the documentation how to install the optional dependencies required to run the example scripts.

E.g., `pip install "shipgrav[examples]"`

openjournals/joss-reviews#7358
