### 2024-10-31 3c84a76
for issues #13 and #14: add progress bars, add tqdm dependency, update pip install instructions for optional packages
instructions for including optional packages with pip installation, issue #14

forgot to add tqdm to dependencies!

one more place that needs tqdm dependency on the list
