+++ shipgrav/__init__.py
@@ -9,9 +9,9 @@ DGS gravimeters output two types of files: serial, or 'raw' files; and 'laptop' 
 Installation
 ------------
 
-shipgrav can be installed from `PyPI <https://pypi.org/project/shipgrav/>`_ using ``pip``. we recommend using an environment management tool like `conda <https://anaconda.org>`_. An exemplary set of commands to make a conda enviroment with shipgrav would be: ::
+shipgrav can be installed from `PyPI <https://pypi.org/project/shipgrav/>`_ using ``pip``. We recommend using an environment management tool like `conda <https://anaconda.org>`_. An exemplary set of commands to make a conda enviroment with shipgrav would be: ::
 
-    conda create --name shipgrav numpy scipy pandas statsmodels tomli pyyaml matplotlib geographiclib
+    conda create --name shipgrav numpy scipy pandas statsmodels tomli pyyaml tqdm matplotlib geographiclib
     conda activate shipgrav
     pip install shipgrav
 
@@ -24,8 +24,11 @@ shipgrav's dependencies are
 * statsmodels
 * tomli
 * pyyaml
+* tqdm
 * matplotlib (optional, needed to run some of the example scripts)
 * geographiclib (optional, needed to run one of the example scripts)
+
+If you install shipgrav with ``pip``, it will also install any of the required dependencies that are missing. To make ``pip`` include the optional dependencies, run ``pip install shipgrav[examples]``.
 
 The example scripts are `available on github <https://github.com/PFPE/shipgrav>`_. They are not packaged with the PyPI package and must be downloaded separately.
 
