+++ docs/requirements.txt
@@ -8,3 +8,4 @@ scipy
 tomli
 statsmodels
 pyyaml
+tqdm
