+++ .github/test_conda_env.yml
@@ -9,3 +9,4 @@ dependencies:
   - tomli
   - pyyaml
   - coverage
+  - tqdm
