### 2024-10-31 PFPE/shipgrav#14 @malmans2
Nice, thanks!
I noticed from the commits that you're suggesting the conda installation and listing dependencies to install. Have you thought about releasing `shipgrav` on `conda-forge` as well? This way, users could simply run `conda install -c conda-forge shipgrav`. Setting up a `conda-forge` recipe is straightforward, and once it's set up, a bot manages most updates for you.
