### 2024-10-31 0599435
update example scripts to include data download via pooch
