### 2024-10-31 f2d065b
add pooch to dependencies, update documentation
