+++ pyproject.toml
@@ -26,7 +26,8 @@ dependencies = [
     "tomli",
     "statsmodels",
     "pyyaml",
-    "tqdm"
+    "tqdm",
+    "pooch"
 ]
 
 [project.optional-dependencies]
