### 2024-10-31 39b44c8
remove the shell script, no longer needed
