### 2024-10-31 553403d
format for jupytext with % syntax, issue #13
