+++ example-scripts/mru_coherence.py
@@ -1,3 +1,21 @@
+# %% [markdown]
+# ### Example script for reading DGS laptop data and MRUs from an R/V Ride transit, and calculating coherence between MRUs and monitors
+#
+# Data files are downloaded by the script using pooch
+#
+# Read DGS laptop and navigation files
+#
+# Correct for meter bias with info from shipgrav
+#
+# Use timestamps to sync more accurate nav with the gravity data.
+#
+# Calculate FAA (free air anomaly) for laptop data
+#
+# Read MRUs (pitch/roll/heave) for all pairs of monitors and MRUs, interpolate MRU to grav sample rate and calculate coherence between monitor and 1000-pt moving average of MRU
+#
+# (and plot coherence)
+
+# %%
 import os
 from glob import glob
 
@@ -13,23 +31,7 @@ from tqdm import tqdm
 import shipgrav.grav as sgg
 import shipgrav.io as sgi
 
-########################################################################
-# Example script for reading DGS laptop data and MRUs from an R/V Ride
-# transit, and calculating coherence between MRUs and monitors
-#
-# Data files are downloaded by the script using pooch
-#
-# Read DGS laptop and navigation files
-# Correct for meter bias with info from shipgrav
-# Use timestamps to sync more accurate nav with the gravity data.
-# Calculate FAA (free air anomaly) for laptop data
-# Read MRUs (pitch/roll/heave)
-# for all pairs of monitors and MRUs, interpolate
-# MRU to grav sample rate and calculate coherence
-# between monitor and 1000-pt moving average of MRU
-# (and plot coherence)
-########################################################################
-
+# %%
 # set some general metadata
 ship = 'Ride'
 cruise = 'SR2312'       # this is used for filepaths
@@ -41,6 +43,7 @@ with open('../shipgrav/database.toml', 'rb') as f:
 nav_tag = info['nav-talkers'][ship]
 biases = info['bias-values'][ship]
 
+# %%
 # get the DGS and nav data from R2R
 dgs_files = pooch.retrieve(url="https://service.rvdata.us/data/cruise/SR2312/fileset/157179", 
         known_hash="53f53c45aa59ce19cd1e75e3d847b5697123ad0a1296aa2be28bf26ff0ad19ac", progressbar=True, 
@@ -60,11 +63,13 @@ nav_files = pooch.retrieve(url="https://service.rvdata.us/data/cruise/SR2312/fil
                      'SR2312/157188/data/SR2312_mru_seapath330_navbho-2023-06-19.txt',
                      'SR2312/157188/data/SR2312_mru_seapath330_navbho-2023-06-20.txt']))
 
+# %%
 # read and sort the nav data
 gps_nav = sgi.read_nav(ship, nav_files, talker='GPGGA')
 gps_nav.sort_values('time_sec', inplace=True)
 gps_nav.reset_index(inplace=True, drop=True)
 
+# %%
 # read and sort the DGS laptop data
 dgs_data = sgi.read_dgs_laptop(dgs_files, ship)
 dgs_data.sort_values('date_time', inplace=True)
@@ -73,6 +78,7 @@ dgs_data['tsec'] = [e.timestamp()
                     for e in dgs_data['date_time']]  # get posix timestamps
 dgs_data['grav'] = dgs_data['rgrav'] + biases['dgs']
 
+# %%
 # sync data geographic coordinates to nav by interpolating with timestamps
 # (interpolators use posix timestamps, not datetimes)
[…]
