+++ example-scripts/dgs_raw_comp.py
@@ -1,3 +1,30 @@
+# %% [markdown]
+# ### Example script for reading and calibrating DGS raw (aka serial) data from an R/V Ride transit, with comparison to laptop (lightly processed) files output from the meter.
+#
+# Data files are downloaded by the script using pooch
+#
+# Read DGS (serial and laptop) and navigation files
+#
+# Calibrate the raw counts
+#
+# Correct for meter bias with info from shipgrav
+#
+# Use timestamps to sync more accurate nav with the gravity data.
+#
+# Plot laptop and raw data to compare
+#
+# Calculate FAA (free air anomaly) for laptop data
+#
+# Plot laptop FAA along with satellite
+#
+# satellite data was tracked from v32.1 Global Gravity grid, which
+# includes data from SIO, NOAA, and NGA.
+#
+# Reference: Sandwell et al. (2014) New global marine gravity model
+# from CryoSat-2 and Jason-1 reveals buried tectonic struture.
+# Science 346(6205), DOI: 10.1126/science.1258213
+
+# %%
 import os
 from glob import glob
 
@@ -13,27 +40,7 @@ import shipgrav.grav as sgg
 import shipgrav.io as sgi
 import shipgrav.utils as sgu
 
-########################################################################
-# Example script for reading and calibrating DGS raw (aka serial) data
-# from an R/V Ride transit, with comparison to laptop (lightly processed)
-# files output from the meter.
-#
-# Data files are downloaded by the script using pooch
-#
-# Read DGS (serial and laptop) and navigation files
-# Calibrate the raw counts
-# Correct for meter bias with info from shipgrav
-# Use timestamps to sync more accurate nav with the gravity data.
-# Plot laptop and raw data to compare
-# Calculate FAA (free air anomaly) for laptop data
-# Plot laptop FAA along with satellite
-# satellite data tracked from v32.1 Global Gravity grid, which
-# includes data from SIO, NOAA, and NGA.
-# Reference: Sandwell et al. (2014) New global marine gravity model
-# from CryoSat-2 and Jason-1 reveals buried tectonic struture.
-# Science 346(6205), DOI: 10.1126/science.1258213
-########################################################################
-
+# %%
 # set some general metadata
 ship = 'Ride'
 cruise = 'SR2312'       # this is used for filepaths
@@ -46,6 +53,7 @@ nav_tag = info['nav-talkers'][ship]
 biases = info['bias-values'][ship]
 cal_factor = info['dgs-stuff']['calibration_factor']
 
+# %%
 # get the data: raw DGS from zenodo, laptop and nav from R2R
 dgs_files_raw = pooch.retrieve(url="https://zenodo.org/records/12733929/files/data.zip", 
         known_hash="md5:83b0411926c0fef9d7ccb2515bb27cc0", progressbar=True, 
@@ -74,17 +82,13 @@ nav_files = pooch.retrieve(url="https://service.rvdata.us/data/cruise/SR2312/fil
                      'SR2312/157188/data/SR2312_mru_seapath330_navbho-2023-06-19.txt',
                      'SR2312/157188/data/SR2312_mru_seapath330_navbho-2023-06-20.txt']))
 
-
-
-# set up file paths, get lists of input files
-root = 'data/'
-dgs_path = os.path.join(root, ship, cruise, 'gravimeter/DGS')
-
+# %%
 # read and sort the nav data
[…]

[…]
