### 2024-10-31 0787008
update optional dependencies for jupyter, issue #13
