### 2024-10-31 PFPE/shipgrav#13 @hfmark
Thanks for all these suggestions - I think I've addressed them, more or less. I didn't fully convert the scripts to notebooks (or embed them in the docs as such) but they are now set up so they can be viewed and run as notebooks using `jupytext`.
