### 2024-10-31 PFPE/shipgrav#10 @hfmark
I can improve this somewhat. Some of what's not covered is i/o functions for the gazillion vessel-specific formats, and that just requires finding some more tiny example files to check against. For the grav module it's a bit harder because the untested functions are (I think) for 2D processing. I added them for completeness, since folks will often look at satellite products alongside trackline data, but I need to think more about how to write good tests for those on very small amounts of data. Anyway, will work on both.
