### 2024-11-01 25a1d87
progress bar flags so they can be turned off for tests
