### 2024-11-01 fdaa515
Revelle BGM timestamps have UTC info already, oops
