### 2024-11-01 0983a1e
add a bunch more tests for better coverage
