+++ shipgrav/tests/test_io.py
@@ -15,35 +15,102 @@ class ioTestCase(unittest.TestCase):
     def ex_files(self):
         return __file__.rsplit(os.sep, 1)[0] + os.sep + "ex_files"
 
-    def test_read_nav(self):  # NOTE only tests one of the ship functions
-        # but it does include extracting clock time -> datetime -> posix
-        nav = sgi.read_nav('Thompson', f'{self.ex_files}'+os.sep+'TN400_nav.Raw')
+    def test_read_nav_Thompson(self):
+        nav = sgi.read_nav('Thompson', f'{self.ex_files}'+os.sep+'TN400_nav.Raw', progressbar=False)
         self.assertEqual(nav.iloc[0].time_sec, 1647129603)
         self.assertTrue(nav.iloc[0].lon + 118.6524 < 0.001)
 
+    def test_read_nav_Atlantis(self):
+        nav = sgi.read_nav('Atlantis', f'{self.ex_files}'+os.sep+'AT01_nav.gps', progressbar=False)
+        self.assertEqual(nav.iloc[0].time_sec, 1656677853)
+        self.assertTrue(nav.iloc[0].lon + 70.67185265 < 0.001)
+
+    def test_read_nav_Langseth(self):
+        nav = sgi.read_nav('Langseth', f'{self.ex_files}'+os.sep+'MGL2003_nav.y2020d244', progressbar=False)
+        self.assertEqual(nav.iloc[0].time_sec, 1598832000.6212)
+        self.assertTrue(nav.iloc[0].lon + 132.620965893 < 0.001)
+
+    def test_read_nav_Revelle(self):
+        nav = sgi.read_nav('Revelle', f'{self.ex_files}'+os.sep+'RR2212_nav.txt', progressbar=False)
+        self.assertEqual(nav.iloc[0].time_sec, 1667606400)
+        self.assertTrue(nav.iloc[0].lon + 119.3889582 < 0.001)
+
+    def test_read_nav_Ride(self):
+        nav = sgi.read_nav('Ride', f'{self.ex_files}'+os.sep+'SR2302_nav.raw',talker='INGGA', progressbar=False)
+        self.assertEqual(nav.iloc[0].time_sec, 1674228947.583)
+        self.assertTrue(nav.iloc[0].lon + 117.23672175 < 0.001)
+
+    def test_read_nav_NBP(self):
+        nav = sgi.read_nav('NBP', f'{self.ex_files}'+os.sep+'NBP_2301_nav.d013', progressbar=False)
+        self.assertEqual(nav.iloc[0].time_sec, 1673568001.81)
+        self.assertTrue(nav.iloc[0].lon + 179.7605851 < 0.001)
+
+    def test_read_nav_nope(self):
+        nav = sgi.read_nav('Titanic', f'{self.ex_files}'+os.sep+'TN400_nav.Raw', progressbar=False)
+        self.assertEqual(nav, -999)
+
     def test_read_bgm_rgs(self):
-        bgm = sgi.read_bgm_rgs(f'{self.ex_files}'+os.sep+'AT05_01_bgm.RGS', 'Atlantis')
+        bgm = sgi.read_bgm_rgs(f'{self.ex_files}'+os.sep+'AT05_01_bgm.RGS', 'Atlantis', progressbar=False)
         self.assertEqual(bgm.iloc[0]['date_time'].timestamp(), 1656633600.445)
         self.assertEqual(bgm.iloc[0]['grav'], 980329.272)
 
-    def test_read_bgm_raw(self):
[…]
