### 2024-11-01 cc38580
better negation for progress bar flag
