### 2024-11-05 PFPE/shipgrav#13 @malmans2
Awesome! Progress bars significantly improve the user experience.
