### 2024-11-05 #7358 @malmans2
Hi all,

I’ve completed my review and checked off all the items on the checklist. I believe `shipgrav` aligns with JOSS standards and is ready for publication. Although I’m not the target user of this library, it looks like a valuable and useful tool for people in the field.
