### 2024-11-05 1504872
pooch is an optional dependency, only needed for examples (PR #15 update)
