### 2024-11-05 8d84428
read for Sikuliaq nav, seapath 320
