### 2024-11-13 PFPE/shipgrav#17 @andreww
Minor documentation issue
Just a very minor documentation glitch while I'm looking. Under installation you say:

> To make pip include the optional dependencies, run `pip install shipgrav[examples]`

You need a space between `shipgrav` and `[examples]` and it could make sense to just say:

> To make pip include the optional dependencies, run 'pip install shipgrav examples'

because it's not optional if you want to add the optional dependencies (if that makes sense).

(Linked to JOSS review at https://github.com/openjournals/joss-reviews/issues/7358)
