### 2024-11-13 PFPE/shipgrav#17 @andreww
Also - and only kind of related - I needed to conda install pooch and tqdm - after pip install shipgrav examples. Maybe this is just pypi being out of sync with github?
