### 2024-11-13 PFPE/shipgrav#17 @andreww
And the final example - RMBA_calc.py - needs geographiclib, same out of sync problem I expect.
