### 2024-11-13 PFPE/shipgrav#18 @andreww
A couple of issues with the examples
If, after installing everything, I run the example `dgs_bgm_comp.py` I get the following output:

![Figure_1](https://github.com/user-attachments/assets/1fa5e549-9d84-42d5-8a00-526aeb326629)

which is not what I expected from looking in the documentation. I've not looked in any detail but wonder if there is something not quite right with the data download for this example?

For `dgs_raw_comp.py` I get the error:

```
python dgs_raw_comp.py
cleaning serial files: 100%|███████████████████████| 5/5 [00:00<00:00, 17360.53it/s]
Traceback (most recent call last):
  File "/Users/andreww/Code/shipgrav/shipgrav/example-scripts/dgs_raw_comp.py", line 157, in <module>
    meter_conf = tm.load(open(opath, 'rb'))
                         ~~~~^^^^^^^^^^^^^
TypeError: expected str, bytes or os.PathLike object, not NoneType
```

which may be as simple as a typo.

All the other examples run fine and give the output shown in the documentation.

(This is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/7358)
