### 2024-11-13 #7358 @andreww
This is so nearly there. My only outstanding issue is that two of the examples aren't giving me the same output as shown in the documentation. The others all work fine. See https://github.com/PFPE/shipgrav/issues/18
