### 2024-11-13 PFPE/shipgrav#17 @hfmark
ah yes, the pooch/tqdm/geographiclib bits are because PyPi is behind github - I just pushed commit 0f53378a16446c746c0222efdba18ac77d4a9fe7 for a new version which should fix those.

For the extra dependencies, I think that `shipgrav[examples]` with the brackets is actually the right syntax? Rather than indicating optional-ness the [] are the cue for "include another set of dependencies that is named in pyproject.toml."
