### 2024-11-13 PFPE/shipgrav#18 @hfmark
That's odd - I can't reproduce either of those errors, even with deleting my pooch cache and re-downloading the data. It's probably worthwhile to try re-pip-installing the new version I just pushed to PyPi and making sure you have the latest example script files pulled from github (though I'm guessing you've done that already), and if that doesn't help, any further info you can dig out would be much appreciated for debugging purposes!
