### 2024-11-20 PFPE/shipgrav#17 @andreww
Ah yes - turns out that I need to escape the square brackets for my shell - `pip install shipgrav\[examples\]` installed it all for me (zsh on a M1 mac mini)
