### 2024-11-20 PFPE/shipgrav#18 @hfmark
Ah, I thought that might be it. I think I found the bug - it's because I didn't explicitly time-sort the BGM data, and I guess different systems end up sorting the two data files in different ways. If the data points aren't ordered then the filtering process goes haywire. I just added a two-line fix and pushed that to the repo - hopefully that will work for you?
