### 2024-11-20 PFPE/shipgrav#17 @hfmark
Good to know! I'll add a note about possibly needing to escape the brackets in the documentation.
