### 2024-11-21 PFPE/shipgrav#18 @andreww
Yes - that fix from yesterday (de9aae5fb40db9a73ae5e46de255add10da2222d) works perfectly for me and fixes this issue.
