### 2024-11-21 #7358 @rwegener2
Thanks again to everyone for their contributions above! @hfmark at this point I'll ask you to please complete the steps in section 2 of [this previous comment](https://github.com/openjournals/joss-reviews/issues/7358#issuecomment-2491713047) including final checks and creating an archive copy of the software. Once that is complete I'll complete my final steps (Section 3) and move the submission back to the Track Editor.
