### 2024-11-21 29bbdd0
bump version v1.0.6 for silly zenodo reasons
