### 2024-11-21 #7358 @hfmark
Wonderful, thank you all! @rwegener2  I've released [v1.0.6](https://github.com/PFPE/shipgrav/releases/tag/v1.0.6) and put it on zenodo (DOI: [10.5281/zenodo.14199589](https://doi.org/10.5281/zenodo.14199589)) with all the authors, orcids, etc confirmed. I think that's all the steps for me for now!
