### 2024-11-21 #7358 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
