### 2024-11-22 #7358 @kthyng
@hfmark You can join JOSS as a reviewer by signing up at https://reviewers.joss.theoj.org/!
